LayoutTests/platform/blackberry/media/video-automatic-fullscreen-expected.txt

EVENT(canplay)
EVENT(webkitfullscreenchange)
END OF TEST
